requirements.txt
numpy~=1.16.2
setuptools~=58.0.4
spidev~=3.5
Pillow~=8.3.2
